requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
openpyxl==3.1.2
reportlab==4.1.0
Werkzeug==3.0.1
